pyproject.toml
[tool.poetry]
name = "photobooth"
version = "0.1.0"
description = "photomaton device"
authors = ["ulric"]

[tool.poetry.dependencies]
python = "^3.7"
pygame = "^2.0.0"
opencv-python = "^4.4.0"
numpy = "1.19.3"
requests = "^2.25.0"
python-networkmanager = "^2.1"
Flask = "^1.1.2"
paramiko = "^2.7.2"

[tool.poetry.dev-dependencies]
pandas = "^1.1.5"
lxml = "^4.6.2"

[tool.poetry.scripts]
photobooth = 'photobooth.photobooth:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
